src/requirements.txt
GitPython==3.1.18
openai==0.8.0
PyYAML==5.4.1